requirements.txt
# Core dependencies
PyYAML>=6.0
httpx>=0.25.0
feedparser>=6.0
Pillow>=10.0

# TTS
edge-tts>=6.1.0

# Speech recognition (optional, for subtitle generation)
# faster-whisper>=1.0.0

# Video composition (system dependency)
# ffmpeg must be installed separately
# espeak-ng must be installed separately

# Development / Testing
pytest>=7.0
pytest-cov>=4.0
